requirements.txt
asgiref==3.4.1
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0
django-phonenumber-field==6.0.0
django-tinymce==3.4.0
mysqlclient==2.1.0
phonenumbers==8.12.39
pip-autoremove==0.10.0
python-decouple==3.5
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
XlsxWriter==3.0.2
